The signal return code was trying to access user mode addresses directly, which doesn't work when the user process is in an entirely different vmspace.  Use copyin() instead.
Assign SIGQUIT (^\) on the console to cause a virtual kernel to enter
its internal db> debugger.

Quickly hack a callout() based poller for console input so the console
works (sorta).

Allow a kernel environment to specified as an option in the form
-e name=value:name=value:...

NULL-out the pmap self-mapping cache when a pmap is destroyed.