Set kernel_vm_end to virtual_start instead of virtual_end so it can be used to detemine how much KVM has been reserved so far (so vm.kvm_free works).
As part of the same deal, redo pmap_growkernel() to just increase
kernel_vm_end up to its limit, instead of just panicing.

Remove in incorrect NULLing out of pmap->pm_pdir.  This was causing the
kernel_map to grow forever until it ran out of space and the system paniced.