Mechanically kill hard sentence breaks.
Fix references to non-existing or obsoleted man pages.

Use ``.Pq Ql'' to quote single characters, instead of ``.Pq Li''.
This makes the resulting output more aesthetically pleasing in
text-only terminals.

mdoc cleanups.

Add missed '.'.

Document that '#' starts a comment.

o Discuss the ways to escape an alias.
o Fix the alias syntax lines.
o Refer to the Aliases subsection.

Document that `in word ...' is optional in the for command.

Document the "return" built-in better: it will exit . (sources) and
the top-level shell instance, too.

When listing the special $ variables, ($!, $#, etc) list them as $!
instead of just !, this allows one to more easily locate/understand
the section of the manpage in question.

Document the -L and -P options to the cd and pwd commands as being
mutually exclusive.

Document command -v and -V.

Split the export synopsis into two lines as -p cannot be mixed with variable
names.

Document the fact that 'true' and 'false' are among sh(1) built-in commands.

- Move the description of the ``-c string'' option closer to the option itself.
- Add an ENVIRONMENT section (1)
- Add more .Xr cross references to the SEE ALSO section.

Add `local' and `return' to the list of built-ins.

Document that read -t timeout returns 1 if the timeout elapses.

Make it easier to find that we have test(1) built-in in sh(1).

- Document trap's -l option and the behaviour of a missing action or a single
  dash.
- Discourage the omission of the action.

Document umask's -S option.

Tell more of the sh(1) history.
Acknowledge Kenneth Almquist's contribution in AUTHORS.

Document missing multibyte character handling in utilities specified
by POSIX.

Obtained-from:  FreeBSD