Do not lookup symbols with STB_LOCAL binding
When a shared object with static TLS symbols gets linked, ld issues
a R_386_TLS_DTPMAP32 relocation on i386, referencing symbol 0, the
undefined symbol.  This is perfectly legal, as this is simply a
local symbol and nothing but the local module's TLS index is needed.

This fixes execution of binaries linked to gcc41's libstdc++.

Obtained-from: FreeBSD