Change kinfo_proc interface between kernel and userland.
Before, we were embedding a struct proc (among others) into struct
kinfo_proc.  Every time we change implementation details in the kernel,
userland has to be adapted (recompiled).  In preparation for the coming
LWP changes this interface has been reworked.  Now kinfo_proc is a
structure which does not depend on other structures on the kernel which
are subject to change.  Instead, the routines fill_kinfo_proc and
fill_kinfo_lwp copy all values which are of interest between the kernel
structure and the stable kinfo_proc structure.

Furthermore, this change adds infrastructure to export LWP-specific data.
If userland requests LWP data, it sets the flag KERN_PROC_FLAG_LWP in the
sysctl oid.  This leads to multiple kinfo_procs being exported.  If not
set, the first LWP will used.  This is like FreeBSD do it, and it seems
easy and simple.  Note that userland was not yet adjusted to actually
request LWPs and aggregate this information if necessary.  Besides, the
kernel does not yet have more than one LWP per process anyways.

This introduces a new file, kern/kern_kinfo.c, which is shared between
kernel and libkvm.  This was done to avoid and remove code duplication.
Now kvm_getprocs constructs a complete struct proc, including pointers,
and then calls fill_kinfo_proc to do its job.

In-collaboration-with: Thomas E. Spanjaard <tgen@netphreax.net>