Update the syslink documentation.  This is still a work in progress.  The main change is that I got stuck in a hole and diverged from my original concept of having a self-identifying network.  These changes reformulate the algorithms such that we are back to the self-identifying network concept. That is, the idea that you can just connect into a cluster at any point and everything will be worked out automatically.
One big issue still needs to be resolved, and that is how to figure out
an optimal physical route from point A to point C without having to go
through point B, where point B is the registration node used as a rendezvous
for A and C.  It is always possible for A to issue a cluster-wide broadcast
to 'forge' a route to C, or vise-versa, and we will probably implement things
this way initially (and always have it available for emergencies), but it
would not scale as the cluster grows despite the fairly rough granularity
in the resources being advertised (e.g. filesystems rather then files).