Add subr_alist.c.  This is a bitmap allocator that works very similarly to subr_blist.c (swap allocator), but with added considerations.
1. All allocations must be in powers of 2.
2. All allocations will be aligned to the allocation size.
3. No allocation size limit (blist was limited to 32 blocks per allocation)

Like the blist allocator, the alist is arranged in a linear array suitable
for direct mapping onto a storage medium.  A dataspace of 2^31-1 blocks may
be represented.  Approximately 3 bits of kernel memory is used per block.

This allocator will be used by HAMMER and ANVIL (filesystem and filesystem
storage manager), and by syslink route nodes to chop out individual addresses
and subnets.  We may also use this allocator to improve the allocation of
physical memory.