Add a proxy libpthread.
In short, it works like this:  ld uses libpthread.so when linking.  There
the SONAME is set to libpthread.so.0.  This however is a symlink (or maybe
varsymlink) to one of the threading libs.  Voila, threading libs can be
switched at execution time and we won't get breakage because of two different
threading libs being used in the same binary.

Idea-by:  joerg and me.