Store the frequency and cputimer used to initialize a periodic systimer. If the system cputimer changes, periodic systimers will be adjusted at the next timeout.
This allows the cputimer to be changed with only minor glitches, which
in turn will allow us to initialize with the 8254 and then later change
over to the ACPI timer.

Theoretically this allows the cputimer to be changed on the fly, after
the system has booted, but the sysctl is read-only for now.