Reduce the livelock limit from 50Khz to 40Khz.
Change the livelock state machine to a one second test based on
gd_time_seconds instead of ticks to try to make the hysteresis
work better.

Add a debug mode so we can measure the interrupt rate while livelocked,
and automatically enable when booting -v (during the boot sequence only).