Changes to consdev - low level kernel console initialization.
The consdev API was calling make_dev() extremely early in the boot sequence,
but except for a little code in syscons didn't really need the abstraction
to operate the kernel console during boot.

Change the consdev API to no longer require the use the device abstraction
to operate.  This will allow the device ABI (cdev_t) to be converted to
use SYSREF.