Add hardware crypto support for 2x61 part of ral(4).  This kind of hardware can do hardware TKIP, TKIP MIC, WEP40/104 and CCMP.
Registers and global/pairwise key table layout are obtained from Ralink RT61
linux driver.  Their driver does not use pairwise key table, but how to setup
pairwise key table and how to enable pairwise key slot are shown.  BUT That's
still NOT enough to use pairwise key --- driver has to write 1 to SEC_CSR4 to
make the wheel whirl.  It is mentioned in no place, and I had to figure it
out myself ;)

2x61 part has 64 pairwise key slots, only last 60 slots will be used currently.
It is mainly because the RX descriptor does not contain a field to indicate
in which key table the decryption key belongs.