Make the kern.ipc.nmbclusters and kern.ipc.nmbufs sysctls read-only. These may only be set as boot-time variables in /boot/loader.conf.
Adjusting them after the fact w/sysctl will not update the objcache limits
and will cause a ton of confusion because then netstat -m will report
the adjusted limits but the system won't actually be using the adjusted
limits.

objcache mbuf limits can be adjusted dynamically, we just need some code to
do it -> for the future!

Reported-by: Peter Avalos <pavalos@theshell.com>