Fix an overflow in the GPT code, I wasn't allocating enough slice structures. Fix the openmask array declaration, it was declaring too-large an array.
Disallow GPT partitions with invalid spans.

When calculating a virgin disklabel take into account the possibility of
absurdly small GPT or MBR slices that would cause the calculation of
disklabel->d_ncylinders to result in 0.  Reduce the number of heads and
the number of sectors per track until a reasonable cylinder count is
achieved.