Implement an architecture function cpu_mplock_contested() which is called by the LWKT thread scheduler when the only thread(s) it can schedule need the MP lock and the scheduler was unable to acquire the MP lock.
On real systems this function just executes the cpu 'pause' instruction,
and on virtual systems this functions sleeps for a millisecond.

Use umtx_sleep() instead of sigpause() in the virtual kernel's idle loop
to interlock threads scheduled via a signal with the idle loop sleep.
This fixes a race condition that caused the vkernel to stop scheduling
(but there may be more issues, stay tuned).