Increase SMP_MAXCPU to 31.  Can't do 32 (boo hoo!) because spinlocks need a bit in the cpumask.
Add DELAY()'s (usleep()'s) in the AP startup code where we spin on the MP
lock.  This greatly improves startup speed when you specify 31 cpus.

Fix a bug in the kqueue interrupt init.  The kqueue code was not registering
its interrupt soon enough which could cause the timer to stop generating
interrupts.