Because the objcache caches up to two magazines on each cpu some pretty bad degenerate conditions will be hit if the cluster limit is set too small or the magazine size is set too large.
Detect the problem and reduce the magazine size to compensate.  If we hit
the minimum magazine size (16), increase the cluster limit to compensate.
Report the corrections on the console.

We also have the option of stealing magazines from other cpus, or reducing
the magazine size even further to handle extreme cases.

This should solve most of the objcache issues when ncpus is set to 31.